2010-04-29	Ralf Corsépius <ralf.corsepius@rtems.org>
	* librpc/src/rpc/pmap_rmt.c: Use uintptr_t for better 16bit
	compliance.

+++ cpukit/librpc/src/rpc/pmap_rmt.c
@@ -281,10 +281,10 @@ clnt_broadcast(
 	struct rpc_msg msg;
 	struct timeval t, tv;
 	char outbuf[MAX_BROADCAST_SIZE], inbuf[UDPMSGSIZE];
-	static u_int32_t disrupt;
+	static uintptr_t disrupt;
 
 	if (disrupt == 0)
-		disrupt = (u_int32_t)(long)resultsp;
+		disrupt = (uintptr_t) resultsp;
 
 	/*
 	 * initialization: create a socket, a broadcast address, and
